LPS-92663 All the layout model listeners were designed for the resource entity, so adding a clause omitting the versions

+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/model/listener/LayoutModelListener.java
@@ -32,6 +32,10 @@ public class LayoutModelListener extends BaseModelListener<Layout> {
 
 	@Override
 	public void onBeforeRemove(Layout layout) throws ModelListenerException {
+		if (!layout.isHead()) {
+			return;
+		}
+
 		try {
 			_journalArticleLocalService.deleteLayoutArticleReferences(
 				layout.getGroupId(), layout.getUuid());
